afxdp/tx/drain: relocate inline tests to drain_tests.rs
Same playbook. drain.rs: 835 → 721 LOC.
cargo build/test 92/878.
